Return eval errors on non symbol defn identifiers
Function definition identifiers cannot begin with digits. This
change adds a check on defn identifiers to ensure they are symbols.
Without this check, the evaluator permits Carp declarations such as
`(defn 4 [] 4)` which are not callable and result in invalid C.
After this change, the evaluator returns an error when it encounters a
defn declaration that contains a non-symbolic identifier such as `4`.